requirement.txt
Pillow
torch==1.11.0
torchvision